requirements-dev.txt
requests
python-dateutil
beautifulsoup4